hestami_copilot/services/training/src/trainer.py
"""
Historian Training Service

Handles QLoRA fine-tuning for the Historian model.
Implements replay buffer strategy to prevent catastrophic forgetting.
"""

import os
import logging
from datetime import datetime
from pathlib import Path
from typing import Optional

import yaml

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Environment configuration
MODEL_PATH = os.getenv("MODEL_PATH", "/models/base")
ADAPTERS_PATH = os.getenv("ADAPTERS_PATH", "/models/adapters")
TRAINING_DATA_PATH = os.getenv("TRAINING_DATA_PATH", "/training-data")
CONFIG_PATH = os.getenv("CONFIG_PATH", "/app/configs/default.yaml")
DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://hestami:hestami@registry-db:5432/historian_registry")


def load_config(config_path: str) -> dict:
    """Load training configuration from YAML file"""
    with open(config_path) as f:
        return yaml.safe_load(f)


def create_adapter_id() -> str:
    """Generate a unique adapter ID based on timestamp"""
    timestamp = datetime.now().strftime("%Y%m%d-%H%M%S")
    return f"historian-lora-{timestamp}"


def prepare_replay_buffer(config: dict, batch_id: str) -> dict:
    """
    Prepare the replay buffer with proper composition.

    Replay buffer composition (from config):
    - new_items_ratio: Fresh training items from recent decisions
    - historical_items_ratio: Previously successful training items
    - constitutional_items_ratio: Core invariant examples (always included)
    """
    ratios = config.get("replay_buffer", {})
    new_ratio = ratios.get("new_items_ratio", 0.10)
    historical_ratio = ratios.get("historical_items_ratio", 0.40)
    constitutional_ratio = ratios.get("constitutional_items_ratio", 0.50)

    logger.info(f"Replay buffer composition: new={new_ratio}, historical={historical_ratio}, constitutional={constitutional_ratio}")

    # Placeholder: In production, this would:
    # 1. Load new training items from batch_id
    # 2. Sample historical items from past successful training
    # 3. Include all constitutional anchors (core invariants)

    return {
        "new_items": [],
        "historical_items": [],
        "constitutional_items": [],
        "total_items": 0,
    }


def validate_hard_gates(metrics: dict, config: dict) -> tuple[bool, list[str]]:
    """
    Validate training results against hard gates.

    Returns (passed, list of failures)
    """
    gates = config.get("hard_gates", {})
    failures = []

    # Check unsupported assertion rate
    if metrics.get("unsupported_assertion_rate", 1.0) > gates.get("max_unsupported_assertion_rate", 0.01):
        failures.append(f"Unsupported assertion rate {metrics['unsupported_assertion_rate']:.2%} exceeds limit")

    # Check UNKNOWN accuracy
    if metrics.get("unknown_accuracy", 0.0) < gates.get("min_unknown_accuracy", 0.95):
        failures.append(f"UNKNOWN accuracy {metrics['unknown_accuracy']:.2%} below threshold")

    # Check label accuracy
    if metrics.get("label_accuracy", 0.0) < gates.get("min_label_accuracy", 0.90):
        failures.append(f"Label accuracy {metrics['label_accuracy']:.2%} below threshold")

    # Check citation precision
    if metrics.get("citation_precision", 0.0) < gates.get("min_citation_precision", 0.95):
        failures.append(f"Citation precision {metrics['citation_precision']:.2%} below threshold")

    # Check JSON validity
    if metrics.get("json_validity", 0.0) < gates.get("min_json_validity", 1.0):
        failures.append(f"JSON validity {metrics['json_validity']:.2%} below 100%")

    return len(failures) == 0, failures


def run_training(batch_id: str, config_path: Optional[str] = None) -> dict:
    """
    Run a training job for the given batch.

    This is a placeholder implementation.
    In production, this would:
    1. Load and validate training data
    2. Prepare replay buffer
    3. Initialize model with QLoRA
    4. Run training loop
    5. Validate against hard gates
    6. Save adapter if passed
    """
    config = load_config(config_path or CONFIG_PATH)
    adapter_id = create_adapter_id()

    logger.info(f"Starting training job for batch: {batch_id}")
    logger.info(f"Adapter ID: {adapter_id}")
    logger.info(f"Base model: {config.get('base_model')}")

    # Prepare replay buffer
    replay_buffer = prepare_replay_buffer(config, batch_id)
    logger.info(f"Replay buffer prepared: {replay_buffer['total_items']} items")

    # Placeholder metrics
    metrics = {
        "unsupported_assertion_rate": 0.005,
        "unknown_accuracy": 0.97,
        "label_accuracy": 0.92,
        "citation_precision": 0.96,
        "json_validity": 1.0,
        "training_loss": 0.15,
    }

    # Validate hard gates
    passed, failures = validate_hard_gates(metrics, config)

    if passed:
        logger.info("All hard gates passed. Adapter ready for promotion.")
        status = "ready_for_promotion"
    else:
        logger.warning(f"Hard gate failures: {failures}")
        status = "failed"

    return {
        "adapter_id": adapter_id,
        "batch_id": batch_id,
        "status": status,
        "metrics": metrics,
        "failures": failures,
    }


def main():
    """Entry point for the training service"""
    logger.info("Historian Training Service starting...")
    logger.info(f"Model path: {MODEL_PATH}")
    logger.info(f"Adapters path: {ADAPTERS_PATH}")
    logger.info(f"Training data path: {TRAINING_DATA_PATH}")

    # Check for pending training batches
    # In production, this would poll the database for pending batches
    # or respond to a message queue

    logger.info("Training service ready. Waiting for training requests...")

    # Placeholder: Keep the service running
    import time
    while True:
        time.sleep(60)
        logger.info("Training service heartbeat")


if __name__ == "__main__":
    main()
